resolve tag/setTag conflicting method names
git-svn-id: svn://orca.physics.unc.edu/Orca64@11882 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Misc Objects/RunningAverage/ORRunningAverageGroup.h
@@ -29,7 +29,7 @@
     BOOL            verbose;
     NSMutableArray* runningAverages;
     int             windowLength;
-    int             tag;
+    NSUInteger      tag;
     int             groupSize;
     float           triggerValue;
     int             triggerType;
@@ -51,8 +51,8 @@
 - (void)        reset;
 - (void)        addValuesUsingTimer;
 - (void)        addNewValue:(float)aValue toIndex:(int)i;
-- (int)         tag;
-- (void)        setTag:(int)newTag;
+- (NSUInteger)         tag;
+- (void)        setTag:(NSUInteger)newTag;
 - (int)         groupSize;
 - (void)        setGroupSize:(int)a;
 - (void)        setVerbose:(BOOL)b;
